+++ coders/histogram.c
@@ -300,7 +300,7 @@ static MagickBooleanType WriteHISTOGRAMImage(const ImageInfo *image_info,
       {
         y=CastDoubleToLong(ceil((double) histogram_image->rows-scale*
           histogram[x].red-0.5));
-        r=q+y*GetPixelChannels(histogram_image);
+        r=q+y*(ssize_t) GetPixelChannels(histogram_image);
         for ( ; y < (ssize_t) histogram_image->rows; y++)
         {
           SetPixelRed(histogram_image,QuantumRange,r);
@@ -311,7 +311,7 @@ static MagickBooleanType WriteHISTOGRAMImage(const ImageInfo *image_info,
       {
         y=CastDoubleToLong(ceil((double) histogram_image->rows-scale*
           histogram[x].green-0.5));
-        r=q+y*GetPixelChannels(histogram_image);
+        r=q+y*(ssize_t) GetPixelChannels(histogram_image);
         for ( ; y < (ssize_t) histogram_image->rows; y++)
         {
           SetPixelGreen(histogram_image,QuantumRange,r);
@@ -322,7 +322,7 @@ static MagickBooleanType WriteHISTOGRAMImage(const ImageInfo *image_info,
       {
         y=CastDoubleToLong(ceil((double) histogram_image->rows-scale*
           histogram[x].blue-0.5));
-        r=q+y*GetPixelChannels(histogram_image);
+        r=q+y*(ssize_t) GetPixelChannels(histogram_image);
         for ( ; y < (ssize_t) histogram_image->rows; y++)
         {
           SetPixelBlue(histogram_image,QuantumRange,r);
